#213 move NetworkService to NetworkUtil for unnecessary service annotation
